feat: AI (#139)
* feat: Added State

* feat: Added board methods in State

* misc: Prefer size_t over int

* misc: Actually, let's use 8 bits over 64 bits.

* style: Formatted the code

* fix: First running version

* misc: Prevent Vector2D copies

* lint: Resolved lint warning

* misc: Put Board on a diet

Getting closer now to the first working version

* perf: Narrow memory usage and started porting the data to State

* fix: First working version

Well, "working", it goes into the GameScene, but the controllers are next to broken.

* docs: Removed useless comment

* misc: First pass at highlighted cells

* fix: Resolved problem with tweening

* fix: State::move not updating the position

* lint: Resolved max-len warning

* fix: Resolved counter-attacks and infinite recursion.

Now units die, though they don't get deleted from memory, and are still rendered. Gotta fix that.

* feat: Added State::getUnitAt to prevent memory copies

* fix: Make the turn logic in the PlayerController smooooooth

* fix: Added killed vector for when units die

* fix: Update healthBar and healthText on damage

* lint: The Return of the Forbidden Bad Line

* fix: Commanders having no meaningful difference when dying

This also fixes the crash. It was caused by Skill being deleted but when that happens, SkillButton is not removed, thus segfaulting as it tried to access properties of a null pointer.

* feat: Make death animation, then kill

A little rusty, it renders the unit for a single frame.

* misc: Tweak Event Loop to tick and destroy before rendering

This should increase accuracy in animation rendering, instead of rendering the right frame in the next frame

* fix: Cast flag to Uint32 to calm down -Wconversion.

* misc: Render HP to 0 when the unit dies

* fix: Units not counter-attacking after its first one

* fix: Resolved edge case when accessing to the scene

* fix: Update battalionSize when damaging a battalion

Also fixed a really old bug

* misc: Change all `byte`s to `uint16_t`.

No more misleading names, please.

* misc: Changed all missing `byte`s.

* cleanup: Remove a lot of the unused methods and properties

* feat: Progress on skills

* fix: Resolved bug where the cooldowns weren't reducing correctly

* misc: Clean up SkillState

* feat: First version of Minimax, removed resolved TODO

* feat: Minimax

* fix: Resolved build error

* fix: Resolved lint error

* feat: Added UnitState.counterAttackable_

And two more skills! ArrowRain and HeroicStrike

* feat: Added all skills minus one

* feat: Finished all skills and resolved all of their TODOs

* feat: Added passive skills

* feat: Added buff and nerf arrow indicators

* fix: Resolved two TODOs, renderizable -> renderable

* fix: MSVC is seriously shit. Just use GCC.

> Implicitly clones value
> The iterator uses a pointer from the cloned value
> Iterate over the cloned ve... oh wait. Changing stats also changes the modifiers in the range-based loop, and only in MSVC. GG.

* fix: Casting all skills causing the turn to be skipped when the unit can still move

* fix: Skills playing the wrong sounds

* fix: Skills using cooldown instead of duration

* fix: Call setAnimation before mutating

* fix: Infinite loops are infinite annoyance

* fix: HeroicStrike playing the wrong sound

* fix: Call saveStatus() on gameOver victory

Although this fixes a bug, it exposes another, restart comes back with the army as it was last saved.

But this is the wanted behaviour as restart should keep in the same level, thus this would work

* restart fixed

* misc: Change the evaluation board to include health

* fix: Reinforce not updating the battalion size number

* text improvements

+++ Drakhtar/Managers/GameManager.h
@@ -17,8 +17,8 @@ class GameManager final {
   int money_ = BASE_MONEY;
   int level_ = 5;
 
-  std::map<std::string, int>* army_ = nullptr;
-  std::map<std::string, int>* buyingCap_ = nullptr;
+  std::map<std::string, uint16_t>* army_ = nullptr;
+  std::map<std::string, uint16_t>* buyingCap_ = nullptr;
   std::map<UnitType, std::string> typeOrder;
   GameManager();
   ~GameManager();
@@ -30,8 +30,8 @@ class GameManager final {
   void updateUnits(std::vector<Unit*>& units);
 
   int getMoney() const;
-  const std::map<std::string, int>& getArmy() const;
-  const std::map<std::string, int>& getCap() const;
+  const std::map<std::string, uint16_t>& getArmy() const;
+  const std::map<std::string, uint16_t>& getCap() const;
   const std::map<UnitType, std::string>& getTypeOrder() const;
   int getLevel() const;
   void increaseLevel();
@@ -39,5 +39,5 @@ class GameManager final {
   void loseMoney(int money);
   void addMoney(int money);
 
-  void addUnits(const std::string& type, int size) const;
+  void addUnits(const std::string& type, uint16_t size) const;
 };
